fix(title-preview-text): fix preview title wrap behaviour
- use field index text as inline element
- consolidate title preview elements into a single element for clarity
- update shared styles to refactor out preview title styles
- updated usages of field index text across builder fields
