+++ src/client/components/builder/questions/CheckboxField.tsx
@@ -23,7 +23,7 @@ import { useCheckerContext } from '../../../contexts'
 import { BuilderActionEnum, ConfigArrayEnum } from '../../../../util/enums'
 
 import '../../../styles/big-checkbox.css'
-import { FieldIndexText } from './FieldIndexText'
+import { TitlePreviewText } from './TitlePreviewText'
 
 const InputComponent: QuestionFieldComponent = ({ field, index }) => {
   const { title, description } = field
@@ -165,10 +165,7 @@ const PreviewComponent: QuestionFieldComponent = ({ field, index }) => {
   return (
     <VStack sx={commonStyles.fullWidthContainer} spacing={2}>
       <VStack sx={commonStyles.fullWidthContainer} spacing={0}>
-        <HStack>
-          <FieldIndexText index={index} />
-          <Text sx={commonStyles.previewTitle}>{title}</Text>
-        </HStack>
+        <TitlePreviewText index={index}>{title}</TitlePreviewText>
         {description && (
           <Text sx={commonStyles.previewDescription}>{description}</Text>
         )}
