+++ src/client/components/builder/questions/DropdownField.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { BiListUl } from 'react-icons/bi'
 import {
-  HStack,
   VStack,
   Text,
   Input,
@@ -15,7 +14,7 @@ import {
 import { useCheckerContext } from '../../../contexts'
 import { createBuilderField, QuestionFieldComponent } from '../BuilderField'
 import { BuilderActionEnum, ConfigArrayEnum } from '../../../../util/enums'
-import { FieldIndexText } from './FieldIndexText'
+import { TitlePreviewText } from './TitlePreviewText'
 
 const InputComponent: QuestionFieldComponent = ({ field, index }) => {
   const { title, description } = field
@@ -91,10 +90,7 @@ const PreviewComponent: QuestionFieldComponent = ({ field, index }) => {
   return (
     <VStack sx={commonStyles.fullWidthContainer} spacing={3}>
       <VStack sx={commonStyles.fullWidthContainer} spacing={0}>
-        <HStack>
-          <FieldIndexText index={index} />
-          <Text sx={commonStyles.previewTitle}>{title}</Text>
-        </HStack>
+        <TitlePreviewText index={index}>{title}</TitlePreviewText>
         {description && (
           <Text sx={commonStyles.previewDescription}>{description}</Text>
         )}
