+++ src/client/components/builder/questions/NumericField.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { BiHash } from 'react-icons/bi'
 import {
   useStyles,
-  HStack,
   VStack,
   Text,
   Input,
@@ -14,7 +13,7 @@ import {
 import { createBuilderField, QuestionFieldComponent } from '../BuilderField'
 import { useCheckerContext } from '../../../contexts'
 import { BuilderActionEnum, ConfigArrayEnum } from '../../../../util/enums'
-import { FieldIndexText } from './FieldIndexText'
+import { TitlePreviewText } from './TitlePreviewText'
 
 const InputComponent: QuestionFieldComponent = ({ field, index }) => {
   const { title, description } = field
@@ -76,10 +75,7 @@ const PreviewComponent: QuestionFieldComponent = ({ field, index }) => {
   return (
     <VStack sx={commonStyles.fullWidthContainer} spacing={3}>
       <VStack sx={commonStyles.fullWidthContainer} spacing={0}>
-        <HStack>
-          <FieldIndexText index={index} />
-          <Text sx={commonStyles.previewTitle}>{title}</Text>
-        </HStack>
+        <TitlePreviewText index={index}>{title}</TitlePreviewText>
         {description && (
           <Text sx={commonStyles.previewDescription}>{description}</Text>
         )}
