--- src/client/theme/components/builder/questions/FieldIndexText.tsx
@@ -1,7 +1,0 @@
-import { ComponentSingleStyleConfig } from '@chakra-ui/theme'
-
-export const FieldIndexText: ComponentSingleStyleConfig = {
-  baseStyle: {
-    textStyle: 'caption1',
-  },
-}
