+++ src/client/theme/components/builder/questions/TitlePreviewText.tsx
@@ -1,0 +1,14 @@
+import { ComponentSingleStyleConfig } from '@chakra-ui/theme'
+
+export const TitlePreviewText: ComponentSingleStyleConfig = {
+  baseStyle: {
+    title: {
+      textStyle: 'subhead1',
+      fontWeight: 500,
+    },
+    fieldIndex: {
+      textStyle: 'caption1',
+      display: 'inline',
+    },
+  },
+}
