+++ src/client/theme/components/builder/questions/index.ts
@@ -2,12 +2,12 @@ import { TitleField } from './TitleField'
 import { NumericField } from './NumericField'
 import { RadioField } from './RadioField'
 import { CheckboxField } from './CheckboxField'
-import { FieldIndexText } from './FieldIndexText'
+import { TitlePreviewText } from './TitlePreviewText'
 
 export const questions = {
   TitleField,
   NumericField,
   RadioField,
   CheckboxField,
-  FieldIndexText,
+  TitlePreviewText,
 }
